requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy
datetime==5.5
pillow==10.0.0
requests
flask==2.2.2
werkzeug==2.2.3
langchain-upstage
langchain-core
langchain
pydantic==1.*